requirements.txt
streamlit
pandas
scikit-learn
python-docx
pdfplumber
reportlab
matplotlib